apps: move ClusterDeviceSource definition into its own file
